BZR-88 Реализовать слушателя и продьюсера по работе с файлами

+++ src/main/java/org/bazar/app/impl/commands/HandleFileUploadedCommand.java
@@ -2,7 +2,7 @@ package org.bazar.app.impl.commands;
 
 public record HandleFileUploadedCommand(
         String key,
-        Integer size,
+        Long size,
         String contentType
 ) {
 }
